fix: move page components to correct location
- Move page components from src/pages/ to src/components/simulation-viewer/pages/
- Move test files from src/pages/__tests__/ to src/__tests__/pages/
- Update test imports to use absolute paths
- Fix Next.js build error (pages directory is for routes, not components)

All 355 tests passing, build succeeds
