+++ src/components/simulation-viewer/pages/AnalysisBugs.tsx
