+++ src/components/simulation-viewer/pages/EncounterHistory.tsx
